Refatora switch case do PedidoService para PedidoOpcoes
